Fixed text item's bounding rect not being properly updated on resizing. This caused a bug where resizing the text block too fast would cause the text item to get cut off because its bounding rect did not match. changed position and size attribute to use doubles instead of ints.

+++ scdatamodel/positionattribute.cpp
@@ -49,7 +49,7 @@ void PositionAttribute::setValue(const QString position)
         return;
     }
 
-    QPointF nPosition(sl[0].toInt(),sl[1].toInt());
+    QPointF nPosition(sl[0].toDouble(),sl[1].toDouble());
 
     if ( nPosition != _position )
     {
